ordinal changed to long

+++ Entrega 4/sales-online-games/src/main/java/com/sales/online/games/salesonlinegames/Domain/Core/Enuns/GameGenre.java
@@ -15,5 +15,5 @@ public enum GameGenre {
     SPORTS, 
     RACING;
 
-    public final int value = ordinal();
+    public final long value = ordinal();
 }
