+++ Entrega 4/sales-online-games/src/main/java/com/sales/online/games/salesonlinegames/Domain/Core/Enuns/OrderStatus.java
@@ -5,5 +5,5 @@ public enum OrderStatus {
     PROCESSING,
     FINISHED;
 
-    public final int value = ordinal();
+    public final long value = ordinal();
 }
