+++ Entrega 4/sales-online-games/src/main/java/com/sales/online/games/salesonlinegames/Domain/Core/Enuns/PaymentMethod.java
@@ -5,5 +5,5 @@ public enum PaymentMethod {
     DEBIT, 
     CREDIT;
 
-    public final int value = ordinal();
+    public final long value = ordinal();
 }
