+++ Entrega 4/sales-online-games/src/main/java/com/sales/online/games/salesonlinegames/Domain/Core/Enuns/UserType.java
@@ -7,5 +7,5 @@ public enum UserType {
     SALER,
     MANAGER;
 
-    public final int value = ordinal();
+    public final long value = ordinal();
 }
